(QENG-2195) add telly jenkins release pipeline
- standardized gem workflow
