+++ lib/telly/version.rb
@@ -1,0 +1,5 @@
+module Telly
+  module Version
+    STRING = '2.9.0'
+  end
+end
