+++ spec/spec_helper.rb
@@ -1,1 +1,2 @@
 require 'webmock/rspec'
+require 'telly'
